+++ adapter-amazon/src/main/java/com/adchain/adapters/AmazonAdAdapter.java
@@ -50,6 +50,7 @@ public class AmazonAdAdapter extends AdChainAdapter implements AdListener {
 
     @Override
     public void init() {
+        AdRegistration.registerApp(getActivity());
         AdRegistration.enableLogging(isLoggingEnabled());
         AdRegistration.enableTesting(isShowTestAds());
         AdRegistration.setAppKey(amazonAppKey);